pyproject.toml
[project]
name = "trabajo_bases_vectoriales_pontia"
version = "0.1.0"
description = "End2EndFlow"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "fastapi>=0.117.1",
    "google-genai>=1.38.0",
    "grpcio==1.60.1",
    "llama-index>=0.14.2",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-llms-google-genai>=0.5.1",
    "llama-index-vector-stores-qdrant>=0.8.5",
    "qdrant-client==1.15.1",
    "uvicorn>=0.36.0",
]
